pyproject.toml
[project]
name = "translator"
version = "0.1.0"
description = ""
authors = [
    {name = "SwitDuck",email = "BaD4epT@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "absl-py (==2.2.2)",
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "astunparse (==1.6.3)",
    "asyncio (==3.4.3)",
    "backports-tarfile (==1.2.0)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.2)",
    "cachetools (==5.5.2)",
    "certifi (==2025.1.31)",
    "charset-normalizer (==3.4.1)",
    "colorama (==0.4.6)",
    "contourpy (==1.3.0)",
    "cycler (==0.12.1)",
    "distlib (==0.3.9)",
    "distro (==1.9.0)",
    "dulwich (==0.22.8)",
    "editdistance (==0.8.1)",
    "efficientnet (==1.0.0)",
    "essential-generators (==1.0)",
    "exceptiongroup (==1.2.2)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "flatbuffers (==25.2.10)",
    "fonttools (==4.57.0)",
    "fsspec (==2025.3.2)",
    "gast (==0.4.0)",
    "google-auth (==2.39.0)",
    "google-auth-oauthlib (==0.4.6)",
    "google-pasta (==0.2.0)",
    "grpcio (==1.71.0)",
    "h11 (==0.14.0)",
    "h5py (==3.13.0)",
    "httpcore (==1.0.8)",
    "httpx (==0.28.1)",
    "huggingface-hub (==0.30.2)",
    "idna (==3.10)",
    "imageio (==2.37.0)",
    "imgaug (==0.4.0)",
    "importlib-metadata (==8.6.1)",
    "importlib-resources (==6.5.2)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "jinja2 (==3.1.6)",
    "jiter (==0.9.0)",
    "keras (==2.11.0)",
    "keras-applications (==1.0.8)",
    "keyring (==25.6.0)",
    "kiwisolver (==1.4.7)",
    "lazy-loader (==0.4)",
    "libclang (==18.1.1)",
    "lxml (==5.4.0)",
    "markdown (==3.8)",
    "markupsafe (==3.0.2)",
    "matplotlib (==3.9.4)",
    "more-itertools (==10.7.0)",
    "mpmath (==1.3.0)",
    "msgpack (==1.1.0)",
    "networkx (==3.2.1)",
    "numpy (==1.26.4)",
    "oauthlib (==3.2.2)",
    "openai (==1.76.0)",
    "opencv-python (==4.11.0.86)",
    "opt-einsum (==3.4.0)",
    "packaging (==25.0)",
    "pbs-installer (==2025.4.9)",
    "pdf2image (==1.17.0)",
    "pillow (==11.2.1)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.7)",
    "protobuf (==3.19.6)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pyclipper (==1.3.0.post6)",
    "pydantic (==2.11.3)",
    "pydantic-core (==2.33.1)",
    "pyparsing (==3.2.3)",
    "pypdf2 (==3.0.1)",
    "pyproject-hooks (==1.2.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-docx (==1.1.2)",
    "pywin32-ctypes (==0.2.3)",
    "pyyaml (==6.0.2)",
    "rapidfuzz (==3.13.0)",
    "regex (==2024.11.6)",
    "requests (==2.32.3)",
    "requests-oauthlib (==2.0.0)",
    "requests-toolbelt (==1.0.0)",
    "safetensors (==0.5.3)",
    "scikit-image (==0.24.0)",
    "scipy (==1.13.1)",
    "sentencepiece (==0.2.0)",
    "shapely (==2.0.7)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sympy (==1.13.3)",
    "tensorboard (==2.11.2)",
    "tensorboard-data-server (==0.6.1)",
    "tensorboard-plugin-wit (==1.8.1)",
    "tensorflow-estimator (==2.11.0)",
    "termcolor (==3.0.1)",
    "tifffile (==2024.8.30)",
    "tokenizers (==0.21.1)",
    "tomli (==2.2.1)",
    "tomlkit (==0.13.2)",
    "torch (==2.7.0)",
    "tqdm (==4.67.1)",
    "transformers (==4.51.3)",
    "trove-classifiers (==2025.4.11.15)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.13.2)",
    "urllib3 (==2.4.0)",
    "validators (==0.34.0)",
    "virtualenv (==20.30.0)",
    "werkzeug (==3.1.3)",
    "wrapt (==1.17.2)",
    "zipp (==3.21.0)",
    "zstandard (==0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
